Split test utilities in different files by concern.
